requirements.txt
Django==4.2.27
django-extensions==4.1
djangorestframework==3.16.1
python-dotenv==1.2.1
black==22.3.0
pylint==2.14.5
autoflake==1.4